+++ Sources/Sentry/include/SentryArrayFilter.h
@@ -1,0 +1,13 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface SentryArrayFilter : NSObject
+
++ (BOOL)isInFuture:(NSDate *_Nullable)date;
+
++ (NSDate *_Nullable)getMaximumDate:(NSDate *_Nullable)first andOther:(NSDate *_Nullable)second;
+
+@end
+
+NS_ASSUME_NONNULL_END